prepare molecules/AlertMessage
