+++ components/sample-form.tsx
@@ -1,19 +1,10 @@
 import React, { ReactElement } from 'react'
-import { useForm, ErrorMessage } from 'react-hook-form'
+import { useForm } from 'react-hook-form'
 import Label from './organisms/sample-form/Label'
 import Password from './organisms/sample-form/Password'
 import Select from './organisms/sample-form/Select'
 import TextLabel from './organisms/sample-form/TextLabel'
-import { Alert } from './atoms/Alert'
-
-type AlertMessageArguments = Readonly<{ errors: string[], name: string}>
-const AlertMessage = ({ errors, name }: AlertMessageArguments): React.ReactElement => {
-  return (<ErrorMessage errors={errors} name={name}>
-    {({ message }: { message: string }): React.ReactElement => (
-      <Alert>{message}</Alert>
-    )}
-  </ErrorMessage>)
-}
+import { AlertMessage } from './molecules/AlertMessage'
 
 const Submit = (): React.ReactElement => {
   return (
